refactored MapCenter and improved supercluster
