+++ src/components/MapCenter/constants.ts
@@ -1,0 +1,1 @@
+export const MAX_ZOOM = 17;
