+++ src/components/MapCenter/hooks/useSupercluster.ts
@@ -1,0 +1,60 @@
+import type { BBox } from 'geojson';
+import { useMemo } from 'react';
+import Supercluster from 'supercluster';
+import type { Business } from '../../../types';
+import { MapService } from '../../../types';
+import { MAX_ZOOM } from '../constants';
+
+interface UseSuperclusterParams {
+	filteredMarkers: Business[];
+	debouncedBounds: BBox | undefined;
+	debouncedZoom: number | undefined;
+	mapService: MapService;
+}
+
+export const useSupercluster = ({
+	filteredMarkers,
+	debouncedBounds,
+	debouncedZoom,
+	mapService,
+}: UseSuperclusterParams) => {
+	const supercluster = useMemo(() => {
+		const instance = new Supercluster({
+			extent: mapService === MapService.GOOGLE ? 256 : 512,
+			radius: import.meta.env.VITE_SUPERCLUSTER_RADIUS ?? 50,
+			maxZoom: mapService === MapService.GOOGLE ? 15 : MAX_ZOOM - 1,
+			minPoints: 2, // Minimum points to form a cluster
+		});
+
+		const points = filteredMarkers.map((business) => ({
+			type: 'Feature',
+			geometry: {
+				type: 'Point',
+				coordinates: [
+					business.yelpData?.coordinates.longitude,
+					business.yelpData?.coordinates.latitude,
+				],
+			},
+			properties: business,
+		})) as Supercluster.PointFeature<Supercluster.AnyProps>[];
+
+		// Load your points into the index
+		instance.load(points);
+
+		return instance;
+	}, [filteredMarkers, mapService]);
+
+	const clusters = useMemo(() => {
+		if (!debouncedBounds || !supercluster || !debouncedZoom) return [];
+
+		const clusters = supercluster.getClusters(
+			debouncedBounds as BBox,
+			Math.floor(debouncedZoom),
+		);
+
+		return clusters;
+	}, [debouncedBounds, debouncedZoom, supercluster]);
+
+	return { supercluster, clusters };
+};
+
