+++ src/main.tsx
@@ -1,13 +1,13 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
-import { APIProvider } from '@vis.gl/react-google-maps';
+import { createSyncStoragePersister } from '@tanstack/query-sync-storage-persister';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { persistQueryClient } from '@tanstack/react-query-persist-client';
-import { createSyncStoragePersister } from '@tanstack/query-sync-storage-persister';
-import MapCenter from './components/MapCenter.tsx';
-import { MapService } from './types/index.ts';
+import { APIProvider } from '@vis.gl/react-google-maps';
 import { MapProvider } from 'react-map-gl';
+import MapCenter from './components/MapCenter/index.tsx';
+import { MapService } from './types/index.ts';
 
 const queryClient = new QueryClient({
 	defaultOptions: {
@@ -39,14 +39,12 @@ const MapContextProvider = ({ children }: { children: React.ReactNode }) => {
 				{children}
 			</APIProvider>
 		);
-	} else if (MAP_SERVICE === MapService.MAPBOX) {
-		return <MapProvider>{children}</MapProvider>;
 	}
 
-	return null;
+	return <MapProvider>{children}</MapProvider>;
 };
 
-createRoot(document.getElementById('root')!).render(
+createRoot(document.getElementById('root')).render(
 	<StrictMode>
 		<QueryClientProvider client={queryClient}>
 			<MapContextProvider>
